[FLINK-17010] Remove --quiet from s3 get operation
The reason this failed with --quiet is that the test redirects stdout to
get the sorted results:

get_complete_result > "${TEST_DATA_DIR}/complete_result"

5f2f7d637d8073b1034f8c1124f52604745936eb removed the --quiet parameter
from the s3 fetch command, which changed what is printed to stdout. The
s3 command is printing which file it is downloading and this log message
also ends up in the "result" output that is being sorted and checked.